+++ rex_xai/evaluation.py
@@ -6,7 +6,7 @@ from scipy.integrate import simpson
 from scipy.signal import periodogram
 from skimage.measure import shannon_entropy
 
-from rex_xai.extraction import Explanation
+from rex_xai.explanation import Explanation
 from rex_xai._utils import get_map_locations
 from rex_xai.mutant import _apply_to_data
 from rex_xai._utils import set_boolean_mask_value, xlogx
