+++ rex_xai/rex.py
@@ -19,7 +19,7 @@ from rex_xai.config import CausalArgs
 from rex_xai.database import update_database
 from rex_xai.evaluation import Evaluation
 from rex_xai.multi_explanation import MultiExplanation
-from rex_xai.extraction import Explanation
+from rex_xai.explanation import Explanation
 from rex_xai.input_data import Data
 from rex_xai.logger import logger
 from rex_xai.onnx import get_prediction_function
